+++ src/extension/inpage-script/nostr.js
@@ -1,4 +1,4 @@
-import NostrProvider from "../ln/nostr";
+import NostrProvider from "../providers/nostr";
 
 if (document) {
   window.nostr = new NostrProvider();
